extract L-R-I into its own file
